afxdp/sharded_neighbor: relocate inline tests to sharded_neighbor_tests.rs
Same playbook. 296-LOC inline test block extracted to sibling
sharded_neighbor_tests.rs.

After: sharded_neighbor.rs drops from 502 → 209 LOC.

cargo build --release: clean. cargo test --release: 878 passed.
